Initial commit.
Our TEST CPS cards recently stopped working on our infrastructure. After
more tests, I realized that some CPS where OK and some made Apache drop
the connection. This seems to be related to this commit [1] which adds a
new test which is failing when Apache tries to check the client
certificate.

For now, I can't tell if it is a bug in the OpenSSL library or a problem
with the ASIP Santé hardware (the CPS card themselves), but it affects
about 40% of the ASIP Santé web of trust.

References:
  1. http://git.openssl.org/gitweb/?p=openssl.git;a=commitdiff;h=5a1e8c67a90aead86ccc2dda324e8f897d1a044d